src/python/grpc-server/pyproject.toml
[tool.poetry]
name = "grpc-server"
version = "0.1.0"
description = ""
authors = ["pecolynx <pecolynx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.60.1"
protobuf = "^4.25.2"
grpcio-tools = "^1.60.1"
grpcio-reflection = "^1.60.1"
grpcio-health-checking = "^1.60.1"
grpcio-testing = "^1.60.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
